+++ payments_payu/provider.py
@@ -564,7 +564,16 @@ class PayuProvider(BasicProvider):
                     print(refunded_price, payment.total)
                     if data["refund"]["status"] == "FINALIZED":
                         payment.message += data["refund"]["reasonDescription"]
-                        if refunded_price == payment.captured_amount:
+                        if refunded_price >= payment.captured_amount:
+                            if refunded_price > payment.captured_amount:
+                                logger.error(
+                                    "refund %s of payment %s has amount greater than the payment's captured_amount: "
+                                    "%f > %f",
+                                    data["refund"].get("refundId", "???"),
+                                    payment.id,
+                                    refunded_price,
+                                    payment.captured_amount,
+                                )
                             payment.change_status(PaymentStatus.REFUNDED)
                         else:
                             payment.captured_amount -= refunded_price