+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 Unreleased
 **********
 * fix multiple deduction of the refund amount from `payment.captured_amount`
+* change statuses of payments refunded with an amount greater than `payment.captured_amount` to `REFUNDED` instead of just deducing `captured_amount`
 
 1.4.0 (2024-04-12)
 ******************